[PIR+CINN]Refactor lower_cinn_fusion_op_pass logic
